usunięcie prewencyjncyh if'ow, przepisanie reservation-widget.js na typescript
