+++ zapisy/webpack_resources/asset-defs.js
@@ -81,7 +81,7 @@ const AssetDefs = {
   // Schedule app
 
   "schedule-reservation-widget": [
-    path.resolve("apps/schedule/assets/reservation-widget.js"),
+    path.resolve("apps/schedule/assets/reservation-widget.ts"),
   ],
   "schedule-fullcalendar": [
     path.resolve("apps/schedule/assets/fullcalendar.js"),
